# src/main/resources/swagger.yaml
openapi: 3.0.2
info:
  title: Microservicio de Stores
  version: v1
paths:
  /stores:
    get:
      summary: Get all stores
      operationId: getStores
      responses:
        '200':
          $ref: '#/components/responses/ListStores'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Create a Store
      operationId: createStore
      requestBody:
        $ref: '#/components/requestBodies/StoreRequestDTO'
      responses:
        '201':
          description: Store created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  '/stores/{storeId}':
    parameters:
      - $ref: '#/components/parameters/storeId'
    get:
      summary: Get a store by id.
      operationId: getStoreById
      responses:
        '200':
          $ref: '#/components/responses/Store'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    delete:
      summary: Delete a store by id.
      operationId: deleteStoreById
      responses:
        '204':
          description: Store created successfully, no content returned.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    put:
      summary: Update a store by id.
      operationId: updateStoreById
      requestBody:
        $ref: '#/components/requestBodies/StoreRequestDTO'
      responses:
        '200':
          $ref: '#/components/responses/Store'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
components:
  requestBodies:
    StoreRequestDTO:
      description: Create a new store
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StoreRequestDTO'

  parameters:
    storeId:
      name: storeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/storeId'
  responses:
    ListStores:
      description: OK.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StoreDTO'
    Store:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StoreDTO'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 400
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Unauthorized:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 401
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Forbidden:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 403
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    NotFound:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    UnexpectedError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
  schemas:
    StoreRequestDTO:
      type: object
      properties:
        name:
          type: string
          example: New York Store
        address:
          $ref: '#/components/schemas/AddressDTO'
    StoreDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        name:
          type: string
          example: New York Store
        address:
          $ref: '#/components/schemas/AddressDTO'
    AddressDTO:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        country:
          type: string
          example: "USA"
    Problem:
      type: object
      properties:
        error:
          type: string
    storeId:
      title: storeId
      description: storeId
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1